code/analysis/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Stephan Druskat <stephan.druskat@dlr.de>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
cffconvert = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
scp_compliance = 'python.scp_compliance:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
